+++ jkeep-jpa/src/main/java/com/ml/jkeep/jpa/system/entity/sys/Role.java
@@ -85,6 +85,6 @@ public class Role {
      * 删除标记
      */
     @Column(name = "d_flag")
-    private Integer dFlag;
+    private Byte dFlag;
 
 }
