+++ jkeep-jpa/src/main/java/com/ml/jkeep/jpa/system/entity/sys/User.java
@@ -108,6 +108,6 @@ public class User {
      * 删除标记
      */
     @Column(name = "d_flag")
-    private Integer dFlag;
+    private Byte dFlag;
 
 }
